Refactor: Extract Toolbar into its own file
